update static plugin so url_path and file_path can be configure separately.
The idea here is that you might want a url of /static/index.html
while index.html might actually live in
/home/foobar/mycontent/for/tiddlyweb/index.html. The previous code
would not allow this. Now it does by adding static_file_path and
static_url_path to the config. Both default to 'static' if not
set in tiddlywebconfig. There is more information in the docstring.