A storage implementation for TiddlyWeb that can read and write to and from multiple other storage implementations.
This means you can store content to a text store and also a sql store,
but then read it from just one of them.

Or you can write content to a reliable store, then have another store which
simply fires a message off to a queue of some kind which can then operate
on the data at its leisure while TiddlyWeb carries on being fast.

The system is flexible enough to allow as many reader and writer storage
implementations as you like, but there are obvious performance concerns with
this: the calls to the stores are done in sequence.

This is highly experimental at this point, but seems to work, test pass
save a problem with revision IDs in the tests.

Note this requires TiddlyWeb 0.9.38 (not yet released).