When using the caching store, we need to make sure we delete the store attribute from the object before caching it.
Otherwise the memcache libraries will try to pickle the
store attribute and this can fail because pickle does not like
to pickle classes or modules.

This problem wasn't showing in the past because the setting of the
store attribute was not being preserved by some stores.

We can safely delete the store attribute in the caching store because
code up the stack will set it again (to point to the caching store).