Add a wserver twanager plugin.
reloader.py does not work on OS X, because it will not allow an
exec() to be called in a process which has several active threads.
Solving this in reloader while still maintaining the simplicity of
doing things as a threaded plugin was not possible.

So instead 'twanager wserver' is available. This peforms the same
role as reloader, but does it by starting one process which then
spawns a 'twanager server' child process. The parent process watches
python code (using the Monitor class from reloader.py) and kills off
the child and starts another when needed.