Update methodhack plugin to avoid a potential bug when POSTing a mocked PUT.
If the content-type is that used for CGI form input, there is a chance
that the code will attempt to read wsgi.input twice. This is not desired.

By moving methodhack before the Query middleware we make sure that REQUEST_METHOD
is PUT before Query is invoked. This means that Query will not read and
process the body of the HTTP request. Downstream handler code will do that.

In order to move methodhack, it has to do it's own query string processing
so it now relies on the cgi package where it would have used tiddlyweb.query
(from Query) in the past.