Make a search method which is resilient in the face of revisions.
This results in some _really_ hairy SQL.

Not yet tested in mysql.

Unlikely that this will be terrifically performant: LIKE queries
with %foo% can't use indexes.

If this SQL works in MySQL can switch to doing a fulltext index
on certain columns and then do a match().

And if that continues to not work there's Whoosh.

Note, while developing and debugging this it became clear that
the mapping done for retrieving a single tiddler is very inefficient,
but this was understood when first putting the sql store together:
it was a demo not a robust solution. It seems to be migrating in the
robust direction. Albeit slowly.