First pass at implementing a search index system using Whoosh.
This replaces /search with a call to code that searches the index.

It calls tiddler_written to write tiddler data to the index.

It adds wsearch and wreindex twanager commands to query or redo
the index from the command line.

The schema is configurable so fields can be indexed and queried.
The query parsing is quite robust.

More docs etc to come after further messing about.