Extend the atom plugin so that it can produce a feed which is a diff between revisions.
If the differ.py module is present and a depth=<count> query string argument
is provided to the Atom URL, the tiddler collection represented by the
URL is delivered a feed with entries for diffs between revisions down to a
depth of count.

This code is the first draft, there has been no refactoring yet. There will
be.

Longer term depth could also deal with time spans.