A start at tiddlywebplugins.templates.
This is a system for managing the packaging of jinja templates
with tiddlyweb plugins. It provides for package-based storage
of default templates with overrides in a local dir.

This is experimental, it's not clear that it will work with
eggs. Not because templates can't be loaded out of eggs (they
can be) but because we need to have multiple packages writing
into the templates dir.