Add permissions handling to wimporter.
The tempory bag that gets created is only visible to the current
user, if there is one.

The bags into which content can be desposite are only those bags
which the current user has write or create.

There's still one more step, which is actually doing the permissions
checking when writing the tiddlers.