Add support for a tasters paramter.
If tasters is true (in config) then we only show tiddlers that
have a database fields of tasters set to true.

This is a fairly specific thing for wiki-data, but will be refactored
into some form of usefulness.

taster and open_fields need to play better together. The idea is
that if you get a taster row from the database, you can show all the data.