First attempt at a challenger that supports openid2.
This makes use of the python-openid library which has the guts
to deal with complexities like XRDS and inames.

In limited testing this code works with the google federate auth
endpoint, inames from freexri and old standbys like livejournal
and myopenid.com.

However, it has not been tested with openids that have query
parameters in the endpoint (as I didn't have one handy).

The major issue that this work exposes is that using the claimed_id
or identity as the tiddlyweb.usersign can be a bit of a mess: The ids
from inames or google are long messy hashes, and the ones from google are
application specific hashes.

This means, as people have already suggested/predicted, that the
ideal use of openid is a mapping between an openid and a local
username. I hadn't done this before because the desire was to keep
users and the user store as simple as possible. With the advent of
tiddlywebplugins.magicuser for storing user metadata cleaner things
maybe be possible.

More poking required.