Integrate chrjs into elmo.
(this uses a hacked up version of chrjs, for the moment. The bug
the hack fixes has been fixed, but I didn't bother to incorporate it here)

Elmo is a thing created quite a while ago to demonstrate getting
at TiddlyWeb resources with xhr calls. It used raw jquery to make
stuff happen. This comment replaces the jqery with chrjs and you'll
see that it was a dead easy change. I'm pleased and impressed.

elmo is not particularly suitable for any purpose, but it has been
a good learning tool.