Add a translation layer to magicuser, prior to extract_more_info.
The translation transforms a username from the form extracted
by an extractor into a canonical centralized form. The mapping
is handled via tiddlers in a bag called "MAPUSER".

If an extractor exacts xfoo.example.com
   If a tiddler with title xfoo.example.com is in MAPUSER
      If that tiddler has the field mapped_user
         then the value of mapped_user is returned as the
         current username

By this method any single canonical username can have many
other usernames mapped to it.

(Note this code provides no mechanism for establishing the
mappings, it is assumed that other code will do that. That
other code is responsible for dealing with ensuring correctness
of the mappings.)