Ensure that whoosher does not send a 500 when there is a lock error.
Sometimes two different processes will want to work on the index at the
same time. In the past this would cause an untrapped LockError
exception. Now that exception is watched for, and a loop is performed,
waiting to get the Lock 5 times. If no lock is ever had, a log message
is written.

Also pep8 cleanups.