First commit of tiddlywebplugins.oom
oom is "One of Many" allowing a filter syntax like:

  oom=title:hello,goodbye,monkey

(A tiddler with title hello OR goodbye OR monkey)

It is similar to mselect but only uses one kind of select. It was
developed to try to find some efficiencies over mselect but it turns
out the inner loop where we get the attributes of the tiddlers is the
expensive part so it's not that much better.

Still useful though, so releasing.