Improve markdown wiki regexp to address issue #16
By using a positive lookbehind assertion for whitespace
we can prevent /FooBar and the like from being treated as
wikilinks when they should be treated as links.

This means that it is sometimes harder to get WikiLinks
where you want them, but the fix is a simple matter of
spacing, whereas the error that is being fixed is harder
to deal with. A reasonable tradeoff.