Adjust store initialization to avoid threading conflicts.
The previous code manipulated the existing config, changing
server_store. This could lead to confusion when two threads were
looking at that chunk of data near the same time.

It turns out, though, that a store doesn't need to think of itself
as the "main store", it just needs to be initialized with the right
data. I'm not sure why I made the code as complicated as I did in
the first place.

This newly simplified version assumes a store will put its config
data in the right place and use config appropriately. This means
that it _won't_ refere back to the server_store part of config
within itself, after initialization.

For the text store, at least, this is true, so we'll see how it goes.